+++ src/generation/gen-data.js
@@ -1,0 +1,29 @@
+function defineGeneratedDataObject()
+{
+	var defineRes = {};
+	
+	defineRes["baseEntries"] = [];
+	defineRes["otherLanguages"] = [];
+	defineRes["checks"] = [];
+	defineRes["personality"] = [];
+	defineRes["hobbies"] = [];
+	defineRes["gaming"] = [];
+	defineRes["avaliability"] = [];
+	defineRes["allergies"] = [];
+	defineRes["pets"] = [];
+	defineRes["fears"] = [];
+	defineRes["technology"] = [];
+	defineRes["qualifications"] = [];
+	defineRes["experienceAreas"] = [];
+	defineRes["previousExperience"] = [];
+	defineRes["other"] = [];
+	
+	return defineRes;
+}
+
+
+
+module.exports =
+{
+	defineObject: defineGeneratedDataObject
+};